backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
sqlalchemy==2.0.23
databases
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT
python-decouple
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0